requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# ── Auth ─────────────────────────────────────────────────────────────────────
bcrypt==4.2.0
PyJWT==2.9.0
pydantic[email]==2.9.2

# ── Task Queue ────────────────────────────────────────────────────────────────
celery==5.4.0
redis==5.1.1

# ── Vector Store ─────────────────────────────────────────────────────────────
chromadb==0.5.18

# ── PDF Processing ────────────────────────────────────────────────────────────
PyMuPDF==1.24.11

# ── ML / Embeddings ──────────────────────────────────────────────────────────
sentence-transformers==3.1.1
torch==2.4.1
transformers==4.45.2
numpy==1.26.4

# ── LLM Providers ────────────────────────────────────────────────────────────
groq==0.11.0
openai==1.51.2
anthropic==0.34.2
google-generativeai==0.8.3

# ── RAG / Search ─────────────────────────────────────────────────────────────
rank-bm25==0.2.2
rapidfuzz==3.10.0

# ── Evaluation ───────────────────────────────────────────────────────────────
ragas==0.2.6
langchain-groq==0.2.0
langchain-huggingface==0.1.0
datasets==3.0.1

# ── Graph / Pipeline ─────────────────────────────────────────────────────────
langgraph==0.2.38
langchain==0.3.4

# ── Utils ────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.2
